+++ database/migrations/1683908983621_lot_produits.ts
@@ -5,8 +5,8 @@ export default class extends BaseSchema {
 
   public async up () {
     this.schema.createTable(this.tableName, (table) => {
-      table.increments('id')
-      table.integer('id_produit').notNullable().unsigned().references('id').inTable('produits').onDelete('CASCADE')
+      table.bigIncrements('id')
+      table.bigInteger('id_produit').notNullable().unsigned().references('id').inTable('produits').onDelete('CASCADE')
       table.integer('qte', 11).notNullable()
       table.integer('prix_achat', 11).notNullable()
